+++ diesel/src/expression/sql_literal.rs
@@ -4,8 +4,6 @@ use backend::Backend;
 use expression::*;
 use query_builder::*;
 use result::QueryResult;
-#[cfg(feature = "with-deprecated")]
-use super::unchecked_bind::UncheckedBind;
 use types::HasSqlType;
 
 #[derive(Debug, Clone)]
@@ -28,82 +26,6 @@ impl<ST> SqlLiteral<ST> {
             sql: sql,
             _marker: PhantomData,
         }
-    }
-
-    /// Bind a value for use with this SQL query.
-    ///
-    /// # Safety
-    ///
-    /// This function should be used with care, as Diesel cannot validate that
-    /// the value is of the right type nor can it validate that you have passed
-    /// the correct number of parameters.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # #[macro_use] extern crate diesel;
-    /// # include!("../doctest_setup.rs");
-    /// #
-    /// # table! {
-    /// #     users {
-    /// #         id -> Integer,
-    /// #         name -> VarChar,
-    /// #     }
-    /// # }
-    /// #
-    /// # fn main() {
-    /// #     use self::users::dsl::*;
-    /// #     use diesel::dsl::sql;
-    /// #     use diesel::types::{Integer, Text};
-    /// #     let connection = establish_connection();
-    /// #[cfg(feature="postgres")]
-    /// let query = sql::<Integer>("SELECT id FROM users WHERE name = $1");
-    /// #[cfg(not(feature="postgres"))]
-    /// let query = sql::<Integer>("SELECT id FROM users WHERE name = ?");
-    /// let seans_id = query.clone().bind::<Text, _>("Sean")
-    ///     .get_result(&connection);
-    /// assert_eq!(Ok(1), seans_id);
-    /// let tess_id = query.bind::<Text, _>("Tess")
-    ///     .get_result(&connection);
-    /// assert_eq!(Ok(2), tess_id);
-    /// # }
-    /// ```
-    ///
-    /// ### Multiple Bind Params
-    ///
-    /// ```rust
-    /// # #[macro_use] extern crate diesel;
-    /// # include!("../doctest_setup.rs");
-    /// #
-    /// # table! {
-    /// #     users {
-    /// #         id -> Integer,
-    /// #         name -> VarChar,
-    /// #     }
-    /// # }
-    /// #
-    /// # fn main() {
-    /// #     use self::users::dsl::*;
-    /// #     use diesel::dsl::sql;
-    /// #     use diesel::types::{Integer, Text};
-    /// #     let connection = establish_connection();
-    /// #     diesel::insert_into(users).values(&NewUser::new("Jim"))
-    /// #         .execute(&connection).unwrap();
-    /// #[cfg(not(feature="postgres"))]
-    /// let query = sql::<Text>("SELECT name FROM users WHERE id > ? AND name <> ?");
-    /// #[cfg(feature="postgres")]
-    /// let query = sql("SELECT name FROM users WHERE id > $1 AND name <> $2");
-    /// let query = query
-    ///     .bind::<Integer, _>(1)
-    ///     .bind::<Text, _>("Jim");
-    /// let expected = vec!["Tess".to_string()];
-    /// assert_eq!(Ok(expected), query.load(&connection));
-    /// # }
-    /// ```
-    #[deprecated(since = "0.99.0", note = "use `sql_query` if you need bind parameters")]
-    #[cfg(feature = "with-deprecated")]
-    pub fn bind<BindST, T>(self, bind_value: T) -> UncheckedBind<Self, T, BindST> {
-        UncheckedBind::new(self, bind_value)
     }
 }